# server/doc/openapi.yaml
openapi: 3.0.3
info:
  title: Taipei Metro API
  description: |
    REST API for the Taipei MRT (捷運) transit system.
    Covers lines, stations, and station distances across 9 lines/branches:
    淡水信義 (R), 板南 (BL), 松山新店 (G), 中和新蘆 (O), 文湖 (BR),
    環狀線 (Y), 新北投支線 (R22A), 小碧潭支線 (G03A).
  version: 1.2.0
servers:
  - url: http://localhost:3000/api
    description: Local development server
paths:
...
  /route-time:
    get:
      summary: Find fastest route (minimum travel time) between two stations
      description: |
        Returns the minimum-time route between two stations based on
        precomputed official travel times from the `route_min_time` table.

        The response mirrors `/route` but uses **travel time (minutes)** instead
        of track distance. Each segment shows the line used, boarding/alighting
        stations, and the travel time for that segment.

        蘆洲支線 (orange_luzhou) shares through-service trains with 中和新蘆線,
        so no transfer is counted at 大橋頭 for that branch.
      operationId: getRouteTime
      parameters:
        - name: from_station_id
          in: query
          required: true
          description: Origin station ID
          schema:
            type: integer
            example: 1
        - name: to_station_id
          in: query
          required: true
          description: Destination station ID
          schema:
            type: integer
            example: 29
      responses:
        '200':
          description: Fastest route with line segments and travel times.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RouteTimeResponse'
              examples:
                one_transfer:
                  summary: "淡水(id=1) → 頂埔(id=29)"
                  value:
                    from: { id: 1, code: R28, name: 淡水 }
                    to: { id: 29, code: BL01, name: 頂埔 }
                    total_travel_time_min: 50
                    transfers: 1
                    segments:
                      - line: { id: 1, code: red, name: 淡水信義線, color: '#E3002C' }
                        from: { id: 1, code: R28, name: 淡水 }
                        to: { id: 20, code: R10, name: R台北車站 }
                        travel_time_min: 28
                      - line: { id: 2, code: blue, name: 板南線, color: '#0070BD' }
                        from: { id: 40, code: BL12, name: BL台北車站 }
                        to: { id: 29, code: BL01, name: 頂埔 }
                        travel_time_min: 22
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
  responses:
    BadRequest:
      description: Invalid query parameters.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: "'date' must be a valid date (YYYY-MM-DD) and 'hour' must be 0–23"
    NotFound:
      description: Resource not found.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: Not found
    InternalError:
      description: Internal server error.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: Internal server error
  schemas:
...
    RouteStation:
      type: object
      required: [id, code, name]
      properties:
        id:
          type: integer
          example: 1
        code:
          type: string
          example: R28
        name:
          type: string
          example: 淡水
    RouteLine:
      type: object
      required: [id, code, name, color]
      properties:
        id:
          type: integer
          example: 1
        code:
          type: string
          example: red
        name:
          type: string
          example: 淡水信義線
        color:
          type: string
          example: '#E3002C'
...
    RouteTimeSegment:
      type: object
      required: [line, from, to, travel_time_min]
      properties:
        line:
          $ref: '#/components/schemas/RouteLine'
        from:
          $ref: '#/components/schemas/RouteStation'
        to:
          $ref: '#/components/schemas/RouteStation'
        travel_time_min:
          type: integer
          description: Travel time in minutes for this segment
          example: 28
    RouteTimeResponse:
      type: object
      required: [from, to, total_travel_time_min, transfers, segments]
      properties:
        from:
          $ref: '#/components/schemas/RouteStation'
        to:
          $ref: '#/components/schemas/RouteStation'
        total_travel_time_min:
          type: integer
          description: Total travel time in minutes
          example: 50
        transfers:
          type: integer
          description: Number of line changes (segments − 1)
          example: 1
        segments:
          type: array
          description: One entry per metro line used, in travel order
          items:
            $ref: '#/components/schemas/RouteTimeSegment'
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
          example: Internal server error
...